// src/app/open-case-core/models/enums.ts
export enum RouterOutlets {
    WorkflowMenu = 'WorkflowMenu'
}
export enum MenuButtonClickType {
    ViewFileHistoyDisable = 'ViewFileHistoyDisable',
    ViewFileHistoy = 'ViewFileHistoy',
    WorkflowMenuView = 'WorkflowMenuView',
    WorkflowMenuEdit = 'WorkflowMenuEdit',
    WorkflowRules = 'WorkflowRules',
    ViewProperty = 'ViewProperty',
    SaveMenuEdit = 'SAVE_MENU_EDIT',
    CloseMenuEdit = 'CLOSE_MENU_EDIT',
    FileBaseMenuSave = 'FILE_BASE_MENU_SAVE',
    ViewFormsLibrary = 'VIEW_FORMS_LIBRARY'
}